Moved trigger logic
