+++ Sources/Superwall/Superwall/Config/ConfigLogic.swift
@@ -272,4 +272,13 @@ enum ConfigLogic {
 
     return identifiers
   }
+  
+  static func getTriggersByEventName(from triggers: Set<Trigger>) -> [String: Trigger] {
+    let triggersDictionary = triggers.reduce([String: Trigger]()) { result, trigger in
+      var result = result
+      result[trigger.eventName] = trigger
+      return result
+    }
+    return triggersDictionary
+  }
 }
