+++ Sources/Superwall/Superwall/Session Events/App Session/AppSessionManager.swift
