+++ Tests/SuperwallTests/Superwall/Config/ConfigLogicTests.swift
@@ -845,4 +845,19 @@ final class ConfigLogicTests: XCTestCase {
     )
     XCTAssertEqual(ids, [paywallId1])
   }
+
+  func test_getTriggerDictionary() {
+    let firstTrigger: Trigger = .stub()
+      .setting(\.eventName, to: "abc")
+
+    let secondTrigger: Trigger = .stub()
+      .setting(\.eventName, to: "def")
+
+    let triggers: Set<Trigger> = [
+      firstTrigger, secondTrigger
+    ]
+    let dictionary = ConfigLogic.getTriggersByEventName(from: triggers)
+    XCTAssertEqual(dictionary["abc"], firstTrigger)
+    XCTAssertEqual(dictionary["def"], secondTrigger)
+  }
 }
